Add more classic Rails loader and preloader tests

+++ test/ruby/shared/loader_spec.rb
@@ -26,50 +26,38 @@ class Loader
 		@sockets = {}
 	end
 
+	def self.new_with_sockets(input, output, app_root)
+		result = allocate
+		result.instance_variable_set(:@input, input)
+		result.instance_variable_set(:@output, output)
+		result.instance_variable_set(:@app_root, app_root)
+		result.instance_variable_set(:@sockets, {})
+		return result
+	end
+
 	def close
 		@input.close if !@input.closed?
 		@output.close if !@output.closed?
-		Process.kill('TERM', @pid)
-		begin
-			Process.waitpid(@pid)
-		rescue Errno::ESCHR, Errno::ECHILD
+		if @pid
+			begin
+				Process.kill('TERM', @pid)
+			rescue Errno::ESRCH
+			end
+			begin
+				Process.waitpid(@pid)
+			rescue Errno::ECHILD
+			end
 		end
 	end
 
-	def negotiate_startup(options = {})
-		@input.puts "You have control 1.0"
-		init_message = read_response
+	def start(options = {})
+		init_message = read_response_line
 		if init_message != "I have control 1.0\n"
 			raise "Unknown response initialization message: #{init_message.inspect}"
 		end
-		@input.puts "passenger_root: #{PhusionPassenger.root}"
-		@input.puts "ruby_libdir: #{PhusionPassenger.ruby_libdir}"
-		@input.puts "generation_dir: #{Utils.passenger_tmpdir}"
-		options.each_pair do |key, value|
-			@input.puts "#{key}: #{value}"
-		end
-		@input.puts
-
-		status = read_response
-
-		headers = {}
-		line = read_response
-		while line != "\n"
-			key, value = line.strip.split(/ *: */, 2)
-			if key == "socket"
-				name, address, protocol, concurrency = value.split(';')
-				@sockets[name] = { :address => address, :protocol => protocol, :concurrency => concurrency }
-			else
-				headers[key] = value
-			end
-			line = read_response
-		end
-
-		if status == "Error\n"
-			body = @output.read
-		end
-
-		return { :status => status.strip, :headers => headers, :body => body }
+		write_request_line "You have control 1.0"
+		write_start_request(options)
+		return process_response
 	end
 
 	def connect_and_send_request(options)
@@ -84,20 +72,148 @@ class Loader
 	end
 
 private
-	def read_response
+	def write_request_line(line = "")
+		STDERR.puts "---> #{line}" if DEBUG
+		@input.puts line
+	end
+
+	def read_response_line
 		while true
 			line = @output.readline
+			STDERR.puts "<--- #{line.strip}" if DEBUG
 			if line.start_with?("!> ")
 				line.sub!(/^\!> /, '')
 				return line
 			end
 		end
 	end
+
+	def write_start_request(options)
+		write_request_line "passenger_root: #{PhusionPassenger.root}"
+		write_request_line "ruby_libdir: #{PhusionPassenger.ruby_libdir}"
+		write_request_line "generation_dir: #{Utils.passenger_tmpdir}"
+		write_request_line "log_level: 3" if DEBUG
+		options.each_pair do |key, value|
+			write_request_line "#{key}: #{value}"
+		end
+		write_request_line
+	end
+
+	def process_response
+		status = read_response_line
+
+		headers = {}
+		line = read_response_line
+		while line != "\n"
+			key, value = line.strip.split(/ *: */, 2)
+			if key == "socket"
+				process_socket(value)
+			else
+				headers[key] = value
+			end
+			line = read_response_line
+		end
+
+		if status == "Error\n"
+			body = @output.read
+		end
+
+		return { :status => status.strip, :headers => headers, :body => body }
+	end
+
+	def process_socket(spec)
+		name, address, protocol, concurrency = spec.split(';')
+		@sockets[name] = { :address => address, :protocol => protocol, :concurrency => concurrency }
+	end
+end
+
+class Preloader < Loader
+	def spawn(options = {})
+		socket = Utils.connect_to_server(sockets["spawn"])
+		loader = Loader.new_with_sockets(socket, socket.dup, @app_root)
+		begin
+			loader.send(:write_request_line, "spawn")
+			loader.send(:write_start_request, options)
+			
+			line = loader.output.readline
+			puts "<--- #{line.strip}" if DEBUG
+			if line != "OK\n"
+				raise "Unexpected spawn response status #{line.inspect}"
+			end
+
+			line = loader.output.readline
+			puts "<--- #{line.strip}" if DEBUG
+			loader.instance_variable_set(:@pid, line.to_i)
+
+			return loader
+		rescue
+			loader.close
+			raise
+		end
+	end
+
+private
+	def process_socket(spec)
+		sockets["spawn"] = spec
+	end
+end
+
+module LoaderSpecHelper
+	def self.included(klass)
+		klass.before(:each) do
+			@stubs = []
+		end
+		
+		klass.after(:each) do
+			begin
+				@loader.close if @loader
+				@preloader.close if @preloader
+			ensure
+				@stubs.each do |stub|
+					stub.destroy
+				end
+			end
+		end
+	end
+	
+	def before_start(code)
+		@before_start = code
+	end
+	
+	def after_start(code)
+		@after_start = code
+	end
+	
+	def register_stub(stub)
+		@stubs << stub
+		File.prepend(stub.startup_file, "#{@before_start}\n")
+		File.append(stub.startup_file, "\n#{@after_start}")
+		return stub
+	end
+	
+	def register_app(app)
+		@apps << app
+		return app
+	end
+
+	def perform_request(options)
+		socket = @loader.connect_and_send_request(options)
+		headers = {}
+		line = socket.readline
+		while line != "\r\n"
+			key, value = line.strip.split(/ *: */, 2)
+			headers[key] = value
+			line = socket.readline
+		end
+		body = socket.read
+		socket.close
+		return [headers, body]
+	end
 end
 
 shared_examples_for "a loader" do
 	it "works" do
-		result = start.negotiate_startup
+		result = start
 		result[:status].should == "Ready"
 		headers, body = perform_request(
 			"REQUEST_METHOD" => "GET",
